+++ tests/test_chunked_processing_pipeline.py
@@ -1,20 +1,19 @@
 import os
+
 import numpy as np
 import pandas as pd
-
 from sqlalchemy import Column
 from sqlalchemy.sql.sqltypes import Integer
 
-from datapipe.datatable import DataStore
-from datapipe.compute import build_compute, run_steps, run_changelist
-from datapipe.store.pandas import TableStoreJsonLine
-from datapipe.store.database import TableStoreDB
-from datapipe.compute import Catalog, Pipeline, Table
+from datapipe.compute import (Catalog, Pipeline, Table, build_compute,
+                              run_changelist, run_steps)
 from datapipe.core_steps import BatchTransform, UpdateExternalTable
-from datapipe.types import data_to_index, ChangeList
+from datapipe.datatable import DataStore
+from datapipe.store.database import TableStoreDB
+from datapipe.store.pandas import TableStoreJsonLine
+from datapipe.types import ChangeList, data_to_index
 
 from .util import assert_datatable_equal, assert_df_equal
-
 
 CHUNK_SIZE = 100
 CHUNK_SIZE_SMALL = 3
@@ -33,8 +32,8 @@ TEST_DF = pd.DataFrame(
 
 
 def test_table_store_json_line_reading(tmp_dir, dbconn):
-    def conversion(df):
-        df["y"] = df["x"] ** 2
+    def conversion(df, multiply):
+        df["y"] = df["x"] ** multiply
         return df
 
     x = pd.Series(np.arange(2 * CHUNK_SIZE, dtype=np.int32))
@@ -64,6 +63,9 @@ def test_table_store_json_line_reading(tmp_dir, dbconn):
             inputs=["input_data"],
             outputs=["output_data"],
             chunk_size=CHUNK_SIZE,
+            kwargs=dict(
+                multiply=2,
+            ),
         ),
     ])
 
@@ -93,7 +95,6 @@ def test_transform_with_many_input_and_output_tables(tmp_dir, dbconn):
                 'inp2_data',
                 TEST_SCHEMA,
                 create_table=True,
-            
             )
         ),
         "out1": Table(
@@ -122,7 +123,7 @@ def test_transform_with_many_input_and_output_tables(tmp_dir, dbconn):
             transform,
             inputs=["inp1", "inp2"],
             outputs=["out1", "out2"],
-            chunk_size=CHUNK_SIZE,
+            chunk_size=CHUNK_SIZE
         ),
     ])